images/consumer/requirements.txt
kafka-python-ng
mysql-connector-python
